Enhancements to the standalone dev container.
I've decoupled the standalone dev container from the VS Code one.  This
allows us to better customize the standalone development experience.

As part of this, I've added a parameter to the script that allows the
user to "target" the container to a particular source directory.
Before, it was only possible to edit the contents of the current repo.

I've also renamed `docker-shell.sh` to `dev-shell` to be slightly more
descriptive and concise.