Quite a bit of polishing
Big picture:

I've refocused this repository a bit.  Its purpose was a little nebulous
to me, since it was comprised only of dev container definitions, a few
docs, and some examples -- yet it's called a "toolkit".

To make it more conceptually cohesive, I've re-postured the repo's message
so that the actual "toolkits" are the two dev containers themselves.
These containers deliver a flexible environment with a variety of useful
tools, including some of our own (see the README for the full list).  I've
decided to keep our home-grown tools (specifically `writ` and `pushwasm`)
in separate repos so as not to conflate the purpose of this repo too much.
Also, I feel this approach will allow people to use them in more flexible
ways if they choose, without the need for container at all.

The exception here is the remote debugging tool.  Since it is coupled to
VS Code, I've kept it in this repo, and it is available in the VS Code
container only.

Examples and documentation continue to be provided here, and are fully
buildable within the development containers.

Specifics:

* I've performed a pretty comprehensive makeover of the main README file.
I would like to eventually provide some more detail about the iterative
development workflow, but I've run out of time for now.

* I tidied up and consolidated the `docs` directory.  I moved the Makefile
in there since that was the only thing it was building.  And, and, because
it was only a placeholder, I removed the first quickstart chapter for now.
I also gave back both the quickstart part 2 chapters their "tutorial" titles.
My rationale is that they aren't particularly quick, and it also allows for
a more clear deprecation of the `wasm-udf-tutorial` repo, which the SEs
are currently using quite a lot.  I still think a quickstart guide would
be helpful, but I don't have time to finish writing it right now.

* I've removed the top-level Cargo.toml file.  It was making the
examples take a very long time to compile, and also consume a lot of space.
To this end, I've removed remote debugger integration from most of the
examples except "echo" and "usergenerator".  In the latter case, there
is now a variant the integrates the remote debugger, and one that does
not.  Being able to see the difference seems helpful to me.  Overall, I
think that it is most convenient for the examples to be individually and
quickly compilable.  At the very least, it lets the documentation flow
more smoothly.

* As previously mentioned, I've added the `writ` and `pushwasm` tools to
both containers.

* I've restructured the Dockerfiles for both containers, and was able to
reduce a significant amount of bloat.  The standalone container is now
2GB instead of 3GB.  We can probably do even more here.  Also, I've
consolidated both of the Dockerfiles into their own directory, `/docker`,
for better organization.

* I've pushed an image of the standalone container up to
ghcr.io/singlestore-labs/singlestore-wasm-toolkit-shell, and have made it
public.  The `scripts/dev-shell` script pulls it automatically.  The VS Code
container still requires a full build before use, however.

* In `devcontainer/devcontainer.json`, I cleaned up some comments I felt
might confuse the user, and I added a marker section for mounting
additional directories (this is referenced in the main README file).

* I've modified the documentation to use `cargo wasi build --lib` instead
of `cargo build --target wasm32-wasi`.